Refactored project structure to match the gridsome structure
