+++ public/index.html
@@ -20,13 +20,11 @@
 </head>
 
 <body id="page-top">
+
 <div id="app"></div>
 
-
-<div id="cookie-div"></div>
 <script src="assets/js/jquery.min.js"></script>
 <script src="assets/bootstrap/js/bootstrap.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.4.1/jquery.easing.min.js"></script>
 <script src="assets/js/agency.js"></script>
 <script src="assets/js/purecookie.js"></script>
 <script src="assets/js/main.js"></script>
