+++ src/components/repairs/AppRepairPopups.vue
@@ -1,20 +1,93 @@
 <template>
-  <div class="modal fade portfolio-modal text-center" role="dialog" tabindex="-1" id="guidemodal1">
-    <div class="modal-dialog modal-lg" role="document">
-      <div class="modal-content">
-        <div class="modal-body">
-          <div class="container">
-            <div class="row">
-              <div class="col-lg-8 mx-auto">
-                <div class="modal-body">
-                  <h2 class="text-uppercase">Επισκευη</h2>
-                  <p class="item-intro text-muted">Επισκευή κλαρίνου αλλαγή ελατηρίου - σούστας</p>
-                  <img class="img-fluid d-block mx-auto" src="assets/img/repairs/allagisoustas.jpg">
-                  <p>
-                    Αφαίρεση του σπασμένου ελατηρίου - σούστας από την κολώνα ή το κλειδί και
-                    αλλαγή με νέο ανοξείδωτο ελατήριο.</p>
-                  <button class="btn btn-primary" data-dismiss="modal" type="button"><i
-                      class="fa fa-times"></i><span> Κλείσιμο</span></button>
+  <div>
+    <div class="modal fade portfolio-modal text-center" role="dialog" tabindex="-1" id="guidemodal1">
+      <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-content">
+          <div class="modal-body">
+            <div class="container">
+              <div class="row">
+                <div class="col-lg-8 mx-auto">
+                  <div class="modal-body">
+                    <h2 class="text-uppercase">Επισκευη</h2>
+                    <p class="item-intro text-muted">Επισκευή κλαρίνου αλλαγή ελατηρίου - σούστας</p>
+                    <img class="img-fluid d-block mx-auto" src="assets/img/repairs/allagisoustas.jpg">
+                    <p>
+                      Αφαίρεση του σπασμένου ελατηρίου - σούστας από την κολώνα ή το κλειδί και
+                      αλλαγή με νέο ανοξείδωτο ελατήριο.</p>
+                    <button class="btn btn-primary" data-dismiss="modal" type="button"><i
+                        class="fa fa-times"></i><span> Κλείσιμο</span></button>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="modal fade portfolio-modal text-center" role="dialog" tabindex="-1" id="guidemodal2">
+      <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-content">
+          <div class="modal-body">
+            <div class="container">
+              <div class="row">
+                <div class="col-lg-8 mx-auto">
+                  <div class="modal-body">
+                    <h2 class="text-uppercase">Επισκευη</h2>
+                    <p class="item-intro text-muted">Επισκευή κλαρίνων αλλαγή ταπών</p><img
+                      class="img-fluid d-block mx-auto" src="assets/img/repairs/allagitapwn.jpg">
+                    <p>Αντικατάσταση - αλλαγή των παλιών ταπών του κλαρίνου - κλαρινέτου με καινούργιες
+                      από δέρμα ή μεμβράνη κατά προτίμηση. Υπάρχουν διάφορες ποιότητες ταπών. Γνήσιες
+                      τάπες εργοστασίου, επαγγελματικές, οικονομικές..</p>
+                    <button class="btn btn-primary" data-dismiss="modal" type="button"><i
+                        class="fa fa-times"></i><span> Κλείσιμο</span></button>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="modal fade portfolio-modal text-center" role="dialog" tabindex="-1" id="guidemodal3">
+      <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-content">
+          <div class="modal-body">
+            <div class="container">
+              <div class="row">
+                <div class="col-lg-8 mx-auto">
+                  <div class="modal-body">
+                    <h2 class="text-uppercase">Επισκευη</h2>
+                    <p class="item-intro text-muted">Επισκευή κλαρίνου αλλαγή φελλών</p><img
+                      class="img-fluid d-block mx-auto" src="assets/img/repairs/allagifelwn.gif">
+                    <p>Αλλαγή - αντικατάσταση των κατεστραμένων φελλών στις συνδέσεις του κλαρίνου και
+                      στο επιστόμιο με αυθεντικό φελλό ή με συνθετικό υλικό.
+                    </p>
+                    <ul class="list-unstyled"></ul>
+                    <button class="btn btn-primary" data-dismiss="modal"
+                            type="button"><i class="fa fa-times"></i><span>&nbsp; Κλείσιμο</span></button>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="modal fade portfolio-modal text-center" role="dialog" tabindex="-1" id="guidemodal4">
+      <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-content">
+          <div class="modal-body">
+            <div class="container">
+              <div class="row">
+                <div class="col-lg-8 mx-auto">
+                  <div class="modal-body">
+                    <h2 class="text-uppercase">Επισκευη</h2>
+                    <p class="item-intro text-muted">Επισκευή κλαρίνου, κόλληση σπασμένου κλειδιού</p>
+                    <img class="img-fluid d-block mx-auto" src="assets/img/repairs/kollisikleidiou.jpg">
+                    <p>Κόληση σπασμένου κλειδιού κλαρίνου με ειδική κόληση. Αποκατάσταση κλειδιού.</p>
+                    <button class="btn btn-primary" data-dismiss="modal" type="button"><i
+                        class="fa fa-times"></i><span> Κλείσιμο</span></button>
+                  </div>
                 </div>
               </div>
             </div>
@@ -23,86 +96,6 @@
       </div>
     </div>
   </div>
-  <div class="modal fade portfolio-modal text-center" role="dialog" tabindex="-1" id="guidemodal2">
-    <div class="modal-dialog modal-lg" role="document">
-      <div class="modal-content">
-        <div class="modal-body">
-          <div class="container">
-            <div class="row">
-              <div class="col-lg-8 mx-auto">
-                <div class="modal-body">
-                  <h2 class="text-uppercase">Επισκευη</h2>
-                  <p class="item-intro text-muted">Επισκευή κλαρίνων αλλαγή ταπών</p><img
-                    class="img-fluid d-block mx-auto" src="assets/img/repairs/allagitapwn.jpg">
-                  <p>Αντικατάσταση - αλλαγή των παλιών ταπών του κλαρίνου - κλαρινέτου με καινούργιες
-                    από δέρμα ή μεμβράνη κατά προτίμηση. Υπάρχουν διάφορες ποιότητες ταπών. Γνήσιες
-                    τάπες εργοστασίου, επαγγελματικές, οικονομικές..</p>
-                  <button class="btn btn-primary" data-dismiss="modal" type="button"><i
-                      class="fa fa-times"></i><span> Κλείσιμο</span></button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-  <div class="modal fade portfolio-modal text-center" role="dialog" tabindex="-1" id="guidemodal3">
-    <div class="modal-dialog modal-lg" role="document">
-      <div class="modal-content">
-        <div class="modal-body">
-          <div class="container">
-            <div class="row">
-              <div class="col-lg-8 mx-auto">
-                <div class="modal-body">
-                  <h2 class="text-uppercase">Επισκευη</h2>
-                  <p class="item-intro text-muted">Επισκευή κλαρίνου αλλαγή φελλών</p><img
-                    class="img-fluid d-block mx-auto" src="assets/img/repairs/allagifelwn.gif">
-                  <p>Αλλαγή - αντικατάσταση των κατεστραμένων φελλών στις συνδέσεις του κλαρίνου και
-                    στο επιστόμιο με αυθεντικό φελλό ή με συνθετικό υλικό.
-                  </p>
-                  <ul class="list-unstyled"></ul>
-                  <button class="btn btn-primary" data-dismiss="modal"
-                          type="button"><i class="fa fa-times"></i><span>&nbsp; Κλείσιμο</span></button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-  <div class="modal fade portfolio-modal text-center" role="dialog" tabindex="-1" id="guidemodal4">
-    <div class="modal-dialog modal-lg" role="document">
-      <div class="modal-content">
-        <div class="modal-body">
-          <div class="container">
-            <div class="row">
-              <div class="col-lg-8 mx-auto">
-                <div class="modal-body">
-                  <h2 class="text-uppercase">Επισκευη</h2>
-                  <p class="item-intro text-muted">Επισκευή κλαρίνου, κόλληση σπασμένου κλειδιού</p>
-                  <img class="img-fluid d-block mx-auto" src="assets/img/repairs/kollisikleidiou.jpg">
-                  <p>Κόληση σπασμένου κλειδιού κλαρίνου με ειδική κόληση. Αποκατάσταση κλειδιού.</p>
-                  <button class="btn btn-primary" data-dismiss="modal" type="button"><i
-                      class="fa fa-times"></i><span> Κλείσιμο</span></button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-  <div id="cookie-div"></div>
-  <script src="assets/js/jquery.min.js"></script>
-  <script src="assets/bootstrap/js/bootstrap.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.4.1/jquery.easing.min.js"></script>
-  <script src="assets/js/agency.js"></script>
-  <script src="assets/js/purecookie.js"></script>
-  </body>
-
-  </html>
 </template>
 
 <script>
