--- src/components/tutorials/AppTutorials.vue
@@ -1,72 +1,0 @@
-<template>
-  <section id="portfolio" class="bg-light">
-    <div class="container">
-      <div class="row">
-        <div class="col-lg-12 text-center">
-          <h2 class="text-uppercase section-heading">ΟΔΗΓΙΕΣ</h2>
-          <h3 class="section-subheading text-muted">Οδηγίες πρώτων βοηθειών πνευστών μουσικών οργάνων</h3>
-        </div>
-      </div>
-      <div class="row">
-        <div class="col-sm-6 col-md-4 portfolio-item">
-          <a class="portfolio-link" data-toggle="modal" href="#guidemodal1">
-            <div class="portfolio-hover">
-              <div class="portfolio-hover-content"><i class="fa fa-plus fa-3x"></i></div>
-            </div>
-            <img class="img-fluid" src="assets/img/guides/31_useclarinetregisterkey800x8001.jpg">
-          </a>
-          <div class="portfolio-caption">
-            <p class="text-muted">Οδηγίες για σπασμένο ή χαλασμένο ελατήριο κλαρίνου</p>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-4 portfolio-item">
-          <a class="portfolio-link" data-toggle="modal" href="#guidemodal2">
-            <div class="portfolio-hover">
-              <div class="portfolio-hover-content"><i class="fa fa-plus fa-3x"></i></div>
-            </div>
-            <img class="img-fluid" src="assets/img/guides/32_01.jpg">
-          </a>
-          <div class="portfolio-caption">
-            <p class="text-muted">Αντικατάσταση χαλασμένης τάπας κλαρίνου</p>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-4 portfolio-item">
-          <a class="portfolio-link" data-toggle="modal" href="#guidemodal3">
-            <div class="portfolio-hover">
-              <div class="portfolio-hover-content"><i class="fa fa-plus fa-3x"></i></div>
-            </div>
-            <img class="img-fluid" src="assets/img/guides/33_fixtrumpetmouthpiece800x8001.jpg">
-          </a>
-          <div class="portfolio-caption">
-            <p class="text-muted">Κολλημένο ή σφηνωμένο επιστόμιο τρομπέτας</p>
-          </div>
-        </div>
-        <div class="col-sm-6 col-md-4 portfolio-item">
-          <a class="portfolio-link" data-toggle="modal" href="#guidemodal4">
-            <div class="portfolio-hover">
-              <div class="portfolio-hover-content"><i class="fa fa-plus fa-3x"></i></div>
-            </div>
-            <img class="img-fluid" src="assets/img/guides/34_oiling_valve_500x3751.jpg">
-          </a>
-          <div class="portfolio-caption">
-            <p class="text-muted">Κολλημένη βαλβίδα τρομπέτας</p>
-          </div>
-        </div>
-      </div>
-    </div>
-  </section>
-  <AppTutorialPopups/>
-</template>
-
-<script>
-import AppTutorialPopups from "@/components/tutorials/AppTutorialPopups";
-
-export default {
-  name: "AppTutorials",
-  components: {AppTutorialPopups}
-}
-</script>
-
-<style scoped>
-
-</style>
