+++ src/layouts/AppFooter.vue
@@ -1,0 +1,37 @@
+<template>
+    <footer>
+      <div class="container">
+        <div class="row">
+          <div class="col-md-4"><span class="copyright">Copyright&nbsp;© Art and Music 2022</span></div>
+          <div class="col-md-4">
+            <ul class="list-inline social-buttons">
+              <li class="list-inline-item">
+                <a href="https://www.facebook.com/ArtAndMusicLamprinidisDimitris"><i class="fa fa-facebook"></i></a>
+              </li>
+              <li class="list-inline-item">
+                <a href="mailto:lamprinidis@artandmusic.gr"><i class="far fa-envelope"></i></a>
+              </li>
+              <li class="list-inline-item">
+                <a href="#"><i class="fa fa-linkedin"></i></a>
+              </li>
+            </ul>
+          </div>
+          <div class="col-md-4">
+            <ul class="list-inline quicklinks">
+              <li class="list-inline-item"><a href="/policy" style="color: #353a40">Privacy Policy</a></li>
+              <li class="list-inline-item"><a href="/policy" style="color: #353a40">Terms of Use</a></li>
+            </ul>
+          </div>
+        </div>
+      </div>
+    </footer>
+</template>
+
+
+<script>
+export default {
+  name: "AppFooter"
+}
+</script>
+
+<style scoped></style>
