+++ src/pages/AppMaintenance.vue
@@ -1,0 +1,20 @@
+<template>
+  <AppMaintenanceBody/>
+  <AppMaintenancePopups/>
+
+</template>
+
+<script>
+
+import AppMaintenanceBody from "@/components/maintenance/AppMaintenanceBody";
+import AppMaintenancePopups from "@/components/maintenance/AppMaintenancePopups";
+
+export default {
+  name: "AppMaintenance",
+  components: {AppMaintenancePopups, AppMaintenanceBody}
+}
+</script>
+
+<style scoped>
+
+</style>
