+++ src/pages/AppPolicy.vue
@@ -8,7 +8,7 @@
     <p>We use Your Personal data to provide and improve the Service. By using the Service, You agree to the
       collection
       and use of information in accordance with this Privacy Policy. This Privacy Policy has been created with the
-      help of the <a href="https://www.freeprivacypolicy.com/free-privacy-policy-generator/"
+      help of the <a href="https://www.freeprivacypolicy.com/free-privacy-policy-generator/AppPolicy.vue"
                      target="_blank">Privacy
         Policy Generator</a>.</p>
     <h1>Interpretation and Definitions</h1>
