+++ src/pages/AppRepairs.vue
@@ -1,0 +1,17 @@
+<template>
+  <AppRepairsBody/>
+  <AppRepairPopups/>
+</template>
+
+<script>
+import AppRepairsBody from "@/components/repairs/AppRepairsBody";
+import AppRepairPopups from "@/components/repairs/AppRepairPopups";
+export default {
+  name: "AppRepairs",
+  components: {AppRepairPopups, AppRepairsBody}
+}
+</script>
+
+<style scoped>
+
+</style>
